# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /prediction/generate:
    get:
      operationId: generatePrediction
      tags: [prediction]
      summary: Generate prediction for next draw
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: ["00:01", "13:00", "16:00", "19:00", "22:00", "23:00"]
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
  /prediction/history:
    get:
      operationId: getPredictionHistory
      tags: [prediction]
      summary: Get prediction history
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Prediction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PredictionRecord"
...
components:
  schemas:
...
    EngineScore:
      type: object
      properties:
        name:
          type: string
        prediction:
          type: string
        weight:
          type: number
        score:
          type: number
      required: [name, prediction, weight, score]
    PredictionResult:
      type: object
      properties:
        prediction:
          type: string
        confidence:
          type: number
        period:
          type: string
        engines:
          type: array
          items:
            $ref: "#/components/schemas/EngineScore"
        topCandidates:
          type: array
          items:
            type: object
            properties:
              number:
                type: string
              score:
                type: number
            required: [number, score]
        generatedAt:
          type: string
      required: [prediction, confidence, period, engines, topCandidates, generatedAt]
    PredictionRecord:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        period:
          type: string
        prediction:
          type: string
        actualResult:
          type: ["string", "null"]
        matchedDigits:
          type: ["integer", "null"]
        accuracy:
          type: ["number", "null"]
        status:
          type: string
        createdAt:
          type: string
      required: [id, date, period, prediction, actualResult, matchedDigits, accuracy, status, createdAt]
...